Inline partition helper into the integration test file
get_succeeded_partitions_by_table was the sole caller's only use of the
helper, so pull it out of hub_api.py (shared helper module) and keep it
local to test_hub_backfill.py as _succeeded_partitions_by_table. Imports
_get_auth_headers from hub_api to avoid duplicating the auth logic.
